Use DataElement class
DataElement class
- includes union of int64_t and double, and indication of the type
- does all type casting and comparistion operators
- update code where needed
- change toBinary to receive vector<ElementData>

+++ tests/pass/quantization.cpp
@@ -21,12 +21,12 @@ TEST(quantization, case_conv)
     mv::QuantizationParams weightsQuantParams({120}, {0.00272007}, {0}, {1});
     mv::QuantizationParams biasQuantParams({0}, {2.13339e-05}, {0}, {1});
 
-    std::vector<double> weightsData = mv::utils::generateSequence<double>(64*64);
+    std::vector<int64_t> weightsData = mv::utils::generateSequence<int64_t>(64*64);
     auto weights = om.constant(weightsData, {1, 1, 64, 64}, mv::DType("UInt8"), mv::Order(mv::Order::getColMajorID(4)), "weights");
     weights->set<mv::QuantizationParams>("quantizationParams", weightsQuantParams);
     auto conv = om.conv(input, weights, {1, 1}, {0, 0, 0, 0}, 1);
     auto convOp = om.getSourceOp(conv);
-    std::vector<double> biasesData = {
+    std::vector<int64_t> biasesData = {
          4267,
          14962,
           7493,
